Moved tests to test dir
